requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
tensorflow
pillow
pydantic[email]